requirements.txt
# ===== Core =====
numpy>=1.24.0,<2.0.0
torch>=2.0.0

# ===== Configuration =====
pyyaml>=6.0

# ===== Visualization & Analysis =====
matplotlib>=3.7.0
seaborn>=0.12.0

# ===== Logging =====
tensorboard>=2.13.0

# ===== Testing =====
pytest>=7.3.0
pytest-cov>=4.1.0

# ===== Code Quality =====
ruff>=0.1.0
